requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
curlify==2.2.1
decorator==5.1.1
Faker==13.4.0
idna==3.3
iniconfig==1.1.1
jsonpath-rw==1.4.0
packaging==21.3
pluggy==1.0.0
ply==3.11
py==1.11.0
PyHamcrest==2.0.3
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
